Compilation fix WIP
