+++ project/plugins.sbt
@@ -4,3 +4,4 @@ addSbtPlugin("org.xerial.sbt"    % "sbt-sonatype"       % "3.9.4")
 addSbtPlugin("com.lucidchart"    % "sbt-scalafmt"       % "1.16")
 addSbtPlugin("org.scoverage"     % "sbt-scoverage"      % "1.5.1")
 addSbtPlugin("com.codecommit"    % "sbt-github-actions" % "0.14.2")
+addSbtPlugin("ch.epfl.scala" % "sbt-scala3-migrate" % "0.6.1")
